+++ packages/file-wrapper/CHANGELOG.md
@@ -3,11 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-### [0.3.8-alpha.1](https://github.com/apimatic/apimatic-js-runtime/compare/@apimatic/file-wrapper@0.3.7...@apimatic/file-wrapper@0.3.8-alpha.1) (2025-09-29)
+### [0.3.9](https://github.com/apimatic/apimatic-js-runtime/compare/@apimatic/file-wrapper@0.3.8...@apimatic/file-wrapper@0.3.9) (2025-11-04)
 
-**Note:** Version bump only for package @apimatic/file-wrapper
+### Bug Fixes
 
-### [0.3.8-alpha.0](https://github.com/apimatic/apimatic-js-runtime/compare/@apimatic/file-wrapper@0.3.7...@apimatic/file-wrapper@0.3.8-alpha.0) (2025-09-29)
+- set headers for file and JSON data parts in multipart from request ([#304](https://github.com/apimatic/apimatic-js-runtime/issues/304)) ([10efc46](https://github.com/apimatic/apimatic-js-runtime/commit/10efc469f964a8115070ae6cf366245acb970696))
+
+### [0.3.8](https://github.com/apimatic/apimatic-js-runtime/compare/@apimatic/file-wrapper@0.3.7...@apimatic/file-wrapper@0.3.8) (2025-09-30)
 
 **Note:** Version bump only for package @apimatic/file-wrapper